libs/legacy/pyproject.toml
[tool.poetry]
name = "superagent"
version = "0.0.54"
description = ""
authors = ["Ismail Pelaseyed"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.95.1"}
uvicorn = {extras = ["standard"], version = "^0.22.0"}
gunicorn = "^20.1.0"
openai = "^0.27.6"
requests = "2.29.0"
python-dotenv = "^1.0.0"
python-decouple = "^3.8"
pydantic = "^1.10.7"
sse-starlette = "^1.5.0"
unstructured = "^0.6.4"
bcrypt = "^4.0.1"
pyjwt = "^2.7.0"
ruff = "^0.0.265"
black = "^23.3.0"
uuid = "^1.30"
flake8 = "^6.0.0"
pinecone-client = "^2.2.1"
tiktoken = "^0.4.0"
pypdf = "^3.8.1"
cohere = "^4.5.1"
wolframalpha = "^5.0.0"
huggingface-hub = "^0.15.1"
bs4 = "^0.0.1"
pdfplumber = "^0.9.0"
transformers = "^4.30.2"
youtube-transcript-api = "^0.6.1"
replicate = "^0.8.3"
python-slugify = "^8.0.1"
google-cloud-firestore = "^2.11.1"
google-auth = "^2.21.0"
psychicapi = "^0.8.0"
tabulate = "^0.9.0"
azure-identity = "^1.13.0"
asyncio = "^3.4.3"
colorlog = "^6.7.0"
llama-hub = "^0.0.16"
pygithub = "^1.59.0"
gitpython = "^3.1.32"
vulture = "^2.7"
airbyte-source-stripe = "^3.17.1"
pyairtable = "^2.0.0"
metaphor-python = "^0.1.11"
superagent-py = "^0.0.51"
prisma = "^0.9.1"
openapi-schema-pydantic = "^1.2.4"
langchain = "^0.0.274"
llama-index = "^0.8.13"
langsmith = "^0.0.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.vulture]
exclude = [
    "*settings.py",
    "*/docs/*.py",
    "*/test_*.py",
    "*/.venv/*.py",
    "app/lib/callbacks.py" # Ignoring this file
    ]
ignore_decorators = ["@app.route", "@require_*"]
ignore_names = ["visit_*", "do_*"]
make_whitelist = true
min_confidence = 100
paths = ["."]
sort_by_size = true
verbose = false

[tool.ruff]
exclude = [
    "*settings.py",
    "*/docs/*.py",
    "*/test_*.py",
    "*/.venv/*.py",
    "whitelist.py"
    ]
